requirements.txt
python-telegram-bot[socks]>=20.0
pyyaml>=6.0
httpx[socks]>=0.24.0
